requirements.txt
Jinja2==3.1.1
python-dotenv==0.20.0
xmltodict==0.13.0
